Separated controllers. Implemented toggle sidebar functions.
